fix(optim): preserve Float64 hyperparams in RMSprop/Adagrad GPU paths
Audit item I.14 (Tenzor 1.0 remediation, Stream I — polish).

RMSprop and Adagrad's CUDA/Vulkan fused-step paths cast every Float64
hyperparameter (lr, alpha, eps, weight_decay, momentum, lr_decay) to
float before stuffing it into OpAttributes, losing precision relative
to Adam/AdamW which preserve double via `attrs.set(AttrKey::Lr, lr_)`.

Drop the static_cast<float>(...) wrappers so the dispatch receives the
original Float64.  NewOpAttributes already supports double-valued
entries (see Adam path).  Backends that genuinely need Float32 internal
arithmetic can narrow at the kernel boundary; that's their decision,
not the optimiser's to force.

All 42 RMSprop / Adagrad tests pass across CPU/CUDA/Vulkan/OneAPI/ROCm.

+++ src/nn/optim/adagrad.cpp
@@ -110,10 +110,13 @@ auto Adagrad::step_impl() -> void {
 
             std::vector<Tensor> inputs = {grad_orig, param->tensor(), sum_[i]};
 
+            // Audit item I.14: pass Float64 hyperparams through to dispatch
+            // (was static_cast<float>, losing precision vs Adam/AdamW which
+            // already preserve double via AttrKey::Lr).
             NewOpAttributes attrs;
-            attrs.set(AttrKey::Lr, static_cast<float>(current_lr));
-            attrs.set(AttrKey::Eps, static_cast<float>(eps_));
-            attrs.set(AttrKey::WeightDecay, static_cast<float>(weight_decay_));
+            attrs.set(AttrKey::Lr, current_lr);
+            attrs.set(AttrKey::Eps, eps_);
+            attrs.set(AttrKey::WeightDecay, weight_decay_);
 
             dispatch(OpId::FusedAdagradStep, inputs, attrs);
             continue;
@@ -127,10 +130,10 @@ auto Adagrad::step_impl() -> void {
             std::vector<Tensor> inputs = {param->tensor(), grad_orig, sum_[i]};
 
             NewOpAttributes attrs;
-            attrs.set(AttrKey::Lr, static_cast<float>(current_lr));
-            attrs.set(AttrKey::LrDecay, static_cast<float>(lr_decay_));
-            attrs.set(AttrKey::Eps, static_cast<float>(eps_));
-            attrs.set(AttrKey::WeightDecay, static_cast<float>(weight_decay_));
+            attrs.set(AttrKey::Lr, current_lr);
+            attrs.set(AttrKey::LrDecay, lr_decay_);
+            attrs.set(AttrKey::Eps, eps_);
+            attrs.set(AttrKey::WeightDecay, weight_decay_);
             attrs.set(AttrKey::Step, step_count_);
 
             dispatch(OpId::FusedAdagradStep, inputs, attrs);
